+++ src/lib/components/blocks/BlockquoteBlock.svelte
@@ -24,6 +24,7 @@
 	export const editable = containerApi.editable;
 	export const focusable = containerApi.focusable;
 	export const focus = containerApi.focus;
+	export const parkCaret = containerApi.parkCaret;
 	export const getCursorOffset = containerApi.getCursorOffset;
 	export const getCursorPosition = containerApi.getCursorPosition;
 	export const focusByPath = containerApi.focusByPath;
@@ -39,6 +40,7 @@
 		editable,
 		focusable,
 		focus,
+		parkCaret,
 		getCursorOffset,
 		getCursorPosition,
 		focusByPath,