+++ src/lib/plugins/latex/BlockMath.svelte
@@ -61,6 +61,7 @@
 	export const focusable = true;
 
 	export const focus = leaf.focus;
+	export const parkCaret = leaf.parkCaret;
 	export const focusAtColumn = leaf.focusAtColumn;
 	export const getCursorOffset = leaf.getCursorOffset;
 	export const getSelectedText = leaf.getSelectedText;
@@ -72,6 +73,7 @@
 		editable,
 		focusable,
 		focus,
+		parkCaret,
 		focusAtColumn,
 		getCursorOffset,
 		getSelectedText,